+++ releasenotes.md
@@ -1,8 +1,11 @@
 edge 2020-2-10
 --------------
 - New: added insert between button to tiles
+- New: roughjs.bar and roughjs.pie tiles
+- New: two new templates showing roughjs usage
 - Breaking: removed "title" content from tables.basic
 - Breaking: removed doc.defaultHidden
+- Fix: updated jtree with pivot table fix
 - Fix: when you delete a tile using the button it only deletes that tile
 - Fix: Bigger down arrow for tile dropdown menu
 - Fix: autosave tab after clicking tile picker command